--- a/pkg/ast/format.go
+++ b/pkg/ast/format.go
@@ -194,6 +194,12 @@ func (n *Type) serialize(w io.Writer) {
 }
 
 func fmtType(t *Type) string {
+	var sb strings.Builder
+	fmtExpressionRec(&sb, t, -1)
+	return sb.String()
+}
+
+func fmtEndType(t *Type) string {
 	v := ""
 	switch {
 	case t.Ident != "":
@@ -245,6 +251,45 @@ func fmtInt(i *Int) string {
 	default:
 		return FormatInt(i.Value, i.ValueFmt)
 	}
+}
+
+func fmtExpressionRec(sb *strings.Builder, t *Type, parentPrio int) {
+	if t.Expression == nil {
+		sb.WriteString(fmtEndType(t))
+		return
+	}
+	be := t.Expression
+	myPrio := operatorPrio(be.Operator)
+	parentheses := myPrio < parentPrio
+	if parentheses {
+		sb.WriteByte('(')
+	}
+	fmtExpressionRec(sb, be.Left, myPrio)
+	sb.WriteByte(' ')
+	switch be.Operator {
+	case OperatorCompareEq:
+		sb.WriteString("==")
+	case OperatorCompareNeq:
+		sb.WriteString("!=")
+	case OperatorBinaryAnd:
+		sb.WriteString("&")
+	default:
+		panic(fmt.Sprintf("unknown operator %q", be.Operator))
+	}
+	sb.WriteByte(' ')
+	fmtExpressionRec(sb, be.Right, myPrio)
+	if parentheses {
+		sb.WriteByte(')')
+	}
+}
+
+func operatorPrio(op Operator) int {
+	for _, info := range binaryOperators {
+		if info.op == op {
+			return info.prio
+		}
+	}
+	panic(fmt.Sprintf("unknown operator %q", op))
 }
 
 func comma(i int, or string) string {