--- a/pkg/compiler/check.go
+++ b/pkg/compiler/check.go
@@ -31,7 +31,7 @@ func (comp *compiler) check(consts map[string]uint64) {
 	comp.checkAttributeValues()
 	comp.checkUnused()
 	comp.checkRecursion()
-	comp.checkLenTargets()
+	comp.checkFieldPaths()
 	comp.checkConstructors()
 	comp.checkVarlens()
 	comp.checkDupConsts()
@@ -206,12 +206,21 @@ func (comp *compiler) checkStructFields(n *ast.Struct, typ, name string) {
 		comp.error(n.Pos, "%v %v has no fields, need at least 1 field", typ, name)
 	}
 	hasDirections, hasOutOverlay := false, false
+	prevFieldHadIf := false
 	for fieldIdx, f := range n.Fields {
 		if n.IsUnion {
-			comp.parseAttrs(unionFieldAttrs, f, f.Attrs)
+			_, exprs := comp.parseAttrs(unionFieldAttrs, f, f.Attrs)
+			if fieldIdx > 0 && fieldIdx+1 < len(n.Fields) &&
+				prevFieldHadIf && exprs[attrIf] == nil {
+				comp.error(f.Pos, "either no fields have conditions or all except the last")
+			}
+			prevFieldHadIf = exprs[attrIf] != nil
+			if fieldIdx+1 == len(n.Fields) && exprs[attrIf] != nil {
+				comp.error(f.Pos, "unions must not have if conditions on the last field")
+			}
 			continue
 		}
-		attrs := comp.parseAttrs(structFieldAttrs, f, f.Attrs)
+		attrs, _ := comp.parseAttrs(structFieldAttrs, f, f.Attrs)
 		dirCount := attrs[attrIn] + attrs[attrOut] + attrs[attrInOut]
 		if dirCount != 0 {
 			hasDirections = true
@@ -362,7 +371,7 @@ func (comp *compiler) checkAttributeValues() {
 			}
 		case *ast.Call:
 			attrNames := make(map[string]bool)
-			descAttrs := comp.parseAttrs(callAttrs, n, n.Attrs)
+			descAttrs := comp.parseIntAttrs(callAttrs, n, n.Attrs)
 			for desc := range descAttrs {
 				attrNames[prog.CppName(desc.Name)] = true
 			}
@@ -404,15 +413,16 @@ func (comp *compiler) checkRequiredCallAttrs(call *ast.Call, callAttrNames map[s
 	}
 }
 
-func (comp *compiler) checkLenTargets() {
+func (comp *compiler) checkFieldPaths() {
 	warned := make(map[string]bool)
 	for _, decl := range comp.desc.Nodes {
 		switch n := decl.(type) {
 		case *ast.Call:
 			for _, arg := range n.Args {
 				checked := make(map[string]bool)
+
 				parents := []parentDesc{{fields: n.Args, call: n.Name.Name}}
-				comp.checkLenType(arg.Type, arg.Type, parents, checked, warned, true)
+				comp.checkFieldPathsRec(arg.Type, arg.Type, parents, checked, warned, true)
 			}
 		}
 	}
@@ -444,7 +454,7 @@ func parentTargetName(s *ast.Struct) string {
 	return templateBase(s.Name.Name)
 }
 
-func (comp *compiler) checkLenType(t0, t *ast.Type, parents []parentDesc,
+func (comp *compiler) checkFieldPathsRec(t0, t *ast.Type, parents []parentDesc,
 	checked, warned map[string]bool, isArg bool) {
 	desc := comp.getTypeDesc(t)
 	if desc == typeStruct {
@@ -468,23 +478,40 @@ func (comp *compiler) checkLenType(t0, t *ast.Type, parents []parentDesc,
 		parents = append([]parentDesc{}, parents...)
 		parents = append(parents, parentDesc{name: parentName, fields: fields})
 		for _, fld := range s.Fields {
-			comp.checkLenType(fld.Type, fld.Type, parents, checked, warned, false)
+			comp.checkFieldPathsRec(fld.Type, fld.Type, parents, checked, warned, false)
+			for _, attr := range fld.Attrs {
+				attrDesc := structFieldAttrs[attr.Ident]
+				if attrDesc == nil || attrDesc.Type != exprAttr {
+					continue
+				}
+				if attrDesc == attrIf {
+					comp.checkExprFieldType(fld.Type)
+				}
+				ast.Recursive(func(n ast.Node) bool {
+					exprType, ok := n.(*ast.Type)
+					if !ok || exprType.Ident != valueIdent {
+						return true
+					}
+					comp.validateFieldPath(exprType.Args[0], t0, exprType, parents, warned)
+					return false
+				})(attr.Args[0])
+			}
 		}
-		warned[parentName] = true
 		return
 	}
 	_, args, _ := comp.getArgsBase(t, isArg)
 	for i, arg := range args {
 		argDesc := desc.Args[i]
 		if argDesc.Type == typeArgLenTarget {
-			comp.checkLenTarget(arg, t0, t, parents, warned)
+			comp.validateFieldPath(arg, t0, t, parents, warned)
 		} else if argDesc.Type == typeArgType {
-			comp.checkLenType(t0, arg, parents, checked, warned, argDesc.IsArg)
+			comp.checkFieldPathsRec(t0, arg, parents, checked, warned, argDesc.IsArg)
 		}
 	}
 }
 
-func (comp *compiler) checkLenTarget(arg, t0, t *ast.Type, parents []parentDesc, warned map[string]bool) {
+func (comp *compiler) validateFieldPath(arg, fieldType, t *ast.Type, parents []parentDesc,
+	warned map[string]bool) {
 	targets := append([]*ast.Type{arg}, arg.Colon...)
 	const maxParents = 2
 	for i, target := range targets {
@@ -530,10 +557,10 @@ func (comp *compiler) checkLenTarget(arg, t0, t *ast.Type, parents []parentDesc,
 			parents = parents[:len(parents)-1]
 		}
 	}
-	comp.checkLenTargetRec(t0, t, targets, parents, warned)
+	comp.validateFieldPathRec(fieldType, t, targets, parents, warned)
 }
 
-func (comp *compiler) checkLenTargetRec(t0, t *ast.Type, targets []*ast.Type,
+func (comp *compiler) validateFieldPathRec(t0, t *ast.Type, targets []*ast.Type,
 	parents []parentDesc, warned map[string]bool) {
 	if len(targets) == 0 {
 		if t.Ident == "offsetof" {
@@ -542,6 +569,7 @@ func (comp *compiler) checkLenTargetRec(t0, t *ast.Type, targets []*ast.Type,
 		}
 		return
 	}
+	isValuePath := t.Ident == valueIdent
 	target := targets[0]
 	targets = targets[1:]
 	fields := parents[len(parents)-1].fields
@@ -553,6 +581,9 @@ func (comp *compiler) checkLenTargetRec(t0, t *ast.Type, targets []*ast.Type,
 			comp.error(target.Pos, "%v target %v refers to itself", t.Ident, target.Ident)
 			return
 		}
+		if !comp.checkPathField(target, t, fld) {
+			return
+		}
 		if len(targets) == 0 {
 			if t.Ident == "len" {
 				typ, desc := comp.derefPointers(fld.Type)
@@ -560,11 +591,15 @@ func (comp *compiler) checkLenTargetRec(t0, t *ast.Type, targets []*ast.Type,
 					// We can reach the same struct multiple times starting from different
 					// syscall arguments. Warn only once.
 					if !warned[parents[len(parents)-1].name] {
+						warned[parents[len(parents)-1].name] = true
 						comp.warning(target.Pos, "len target %v refer to an array with"+
 							" variable-size elements (do you mean bytesize?)",
 							target.Ident)
 					}
 				}
+			}
+			if isValuePath {
+				comp.checkExprLastField(target, fld)
 			}
 			return
 		}
@@ -579,7 +614,7 @@ func (comp *compiler) checkLenTargetRec(t0, t *ast.Type, targets []*ast.Type,
 			return
 		}
 		parents = append(parents, parentDesc{name: parentTargetName(s), fields: s.Fields})
-		comp.checkLenTargetRec(t0, t, targets, parents, warned)
+		comp.validateFieldPathRec(t0, t, targets, parents, warned)
 		return
 	}
 	for pi := len(parents) - 1; pi >= 0; pi-- {
@@ -588,7 +623,7 @@ func (comp *compiler) checkLenTargetRec(t0, t *ast.Type, targets []*ast.Type,
 			parent.name == "" && target.Ident == prog.SyscallRef {
 			parents1 := make([]parentDesc, pi+1)
 			copy(parents1, parents[:pi+1])
-			comp.checkLenTargetRec(t0, t, targets, parents1, warned)
+			comp.validateFieldPathRec(t0, t, targets, parents1, warned)
 			return
 		}
 	}
@@ -598,6 +633,33 @@ func (comp *compiler) checkLenTargetRec(t0, t *ast.Type, targets []*ast.Type,
 			t.Ident, target.Ident, parents[len(parents)-1].String())
 	}
 	warned[warnKey] = true
+}
+
+func (comp *compiler) checkPathField(target, t *ast.Type, field *ast.Field) bool {
+	for _, attr := range field.Attrs {
+		desc := structFieldAttrs[attr.Ident]
+		if desc == attrIf {
+			comp.error(target.Pos, "%s has conditions, so %s path cannot reference it",
+				field.Name.Name, t.Ident)
+			return false
+		}
+	}
+
+	return true
+}
+
+func (comp *compiler) checkExprLastField(target *ast.Type, field *ast.Field) {
+	_, desc := comp.derefPointers(field.Type)
+	if desc != typeInt {
+		comp.error(target.Pos, "%v does not refer to an integer", field.Name.Name)
+	}
+}
+
+func (comp *compiler) checkExprFieldType(t *ast.Type) {
+	desc := comp.getTypeDesc(t)
+	if desc == typeInt && len(t.Colon) != 0 {
+		comp.error(t.Pos, "bitfields may not have conditions")
+	}
 }
 
 func CollectUnused(desc *ast.Description, target *targets.Target, eh ast.ErrorHandler) ([]ast.Node, error) {
@@ -818,7 +880,7 @@ func (comp *compiler) checkTypeCtors(t *ast.Type, dir prog.Dir, isArg, canCreate
 		}
 		checked[key] = true
 		for _, fld := range s.Fields {
-			fldDir, fldHasDir := comp.genFieldDir(fld)
+			fldDir, fldHasDir := comp.genFieldDir(comp.parseIntAttrs(structFieldAttrs, fld, fld.Attrs))
 			if !fldHasDir {
 				fldDir = dir
 			}
@@ -1364,21 +1426,22 @@ func (comp *compiler) checkVarlen(n *ast.Struct) {
 	// Non-varlen unions can't have varlen fields.
 	// Non-packed structs can't have varlen fields in the middle.
 	if n.IsUnion {
-		attrs := comp.parseAttrs(unionAttrs, n, n.Attrs)
+		attrs := comp.parseIntAttrs(unionAttrs, n, n.Attrs)
 		if attrs[attrVarlen] != 0 {
 			return
 		}
 	} else {
-		attrs := comp.parseAttrs(structAttrs, n, n.Attrs)
+		attrs := comp.parseIntAttrs(structAttrs, n, n.Attrs)
 		if attrs[attrPacked] != 0 {
 			return
 		}
 	}
 	for i, f := range n.Fields {
+		_, exprs := comp.parseAttrs(structOrUnionFieldAttrs(n), f, f.Attrs)
 		if !n.IsUnion && i == len(n.Fields)-1 {
 			break
 		}
-		if comp.isVarlen(f.Type) {
+		if comp.isVarlen(f.Type) || !n.IsUnion && exprs[attrIf] != nil {
 			if n.IsUnion {
 				comp.error(f.Pos, "variable size field %v in non-varlen union %v",
 					f.Name.Name, n.Name.Name)