--- a/pkg/compiler/compiler.go
+++ b/pkg/compiler/compiler.go
@@ -180,7 +180,7 @@ func (comp *compiler) structIsVarlen(name string) bool {
 	}
 	s := comp.structs[name]
 	if s.IsUnion {
-		res := comp.parseAttrs(unionAttrs, s, s.Attrs)
+		res := comp.parseIntAttrs(unionAttrs, s, s.Attrs)
 		if res[attrVarlen] != 0 {
 			comp.structVarlen[name] = true
 			return true
@@ -189,7 +189,13 @@ func (comp *compiler) structIsVarlen(name string) bool {
 	comp.structVarlen[name] = false // to not hang on recursive types
 	varlen := false
 	for _, fld := range s.Fields {
-		if comp.isVarlen(fld.Type) {
+		hasIfAttr := false
+		for _, attr := range fld.Attrs {
+			if structFieldAttrs[attr.Ident] == attrIf {
+				hasIfAttr = true
+			}
+		}
+		if hasIfAttr || comp.isVarlen(fld.Type) {
 			varlen = true
 			break
 		}
@@ -198,40 +204,71 @@ func (comp *compiler) structIsVarlen(name string) bool {
 	return varlen
 }
 
-func (comp *compiler) parseAttrs(descs map[string]*attrDesc, parent ast.Node, attrs []*ast.Type) map[*attrDesc]uint64 {
+func (comp *compiler) parseIntAttrs(descs map[string]*attrDesc, parent ast.Node,
+	attrs []*ast.Type) map[*attrDesc]uint64 {
+	intAttrs, _ := comp.parseAttrs(descs, parent, attrs)
+	return intAttrs
+}
+
+func (comp *compiler) parseAttrs(descs map[string]*attrDesc, parent ast.Node, attrs []*ast.Type) (
+	map[*attrDesc]uint64, map[*attrDesc]prog.Expression) {
 	_, parentType, parentName := parent.Info()
-	res := make(map[*attrDesc]uint64)
+	resInt := make(map[*attrDesc]uint64)
+	resExpr := make(map[*attrDesc]prog.Expression)
 	for _, attr := range attrs {
 		if unexpected, _, ok := checkTypeKind(attr, kindIdent); !ok {
 			comp.error(attr.Pos, "unexpected %v, expect attribute", unexpected)
-			return res
+			return resInt, resExpr
 		}
 		if len(attr.Colon) != 0 {
 			comp.error(attr.Colon[0].Pos, "unexpected ':'")
-			return res
+			return resInt, resExpr
 		}
 		desc := descs[attr.Ident]
 		if desc == nil {
 			comp.error(attr.Pos, "unknown %v %v attribute %v", parentType, parentName, attr.Ident)
-			return res
+			return resInt, resExpr
 		}
-		if _, ok := res[desc]; ok {
+		_, dupInt := resInt[desc]
+		_, dupExpr := resExpr[desc]
+		if dupInt || dupExpr {
 			comp.error(attr.Pos, "duplicate %v %v attribute %v", parentType, parentName, attr.Ident)
-			return res
+			return resInt, resExpr
 		}
-		val := uint64(1)
-		if desc.HasArg {
-			val = comp.parseAttrArg(attr)
-		} else if len(attr.Args) != 0 {
-			comp.error(attr.Pos, "%v attribute has args", attr.Ident)
-			return res
+
+		switch desc.Type {
+		case flagAttr:
+			resInt[desc] = 1
+			if len(attr.Args) != 0 {
+				comp.error(attr.Pos, "%v attribute has args", attr.Ident)
+				return nil, nil
+			}
+		case intAttr:
+			resInt[desc] = comp.parseAttrIntArg(attr)
+		case exprAttr:
+			resExpr[desc] = comp.parseAttrExprArg(attr)
+		default:
+			comp.error(attr.Pos, "attribute %v has unknown type", attr.Ident)
+			return nil, nil
 		}
-		res[desc] = val
 	}
-	return res
+	return resInt, resExpr
 }
 
-func (comp *compiler) parseAttrArg(attr *ast.Type) uint64 {
+func (comp *compiler) parseAttrExprArg(attr *ast.Type) prog.Expression {
+	if len(attr.Args) != 1 {
+		comp.error(attr.Pos, "%v attribute is expected to have only one argument", attr.Ident)
+		return nil
+	}
+	arg := attr.Args[0]
+	if arg.HasString {
+		comp.error(attr.Pos, "%v argument must be an expression", attr.Ident)
+		return nil
+	}
+	return comp.genExpression(arg)
+}
+
+func (comp *compiler) parseAttrIntArg(attr *ast.Type) uint64 {
 	if len(attr.Args) != 1 {
 		comp.error(attr.Pos, "%v attribute is expected to have 1 argument", attr.Ident)
 		return 0