--- a/sys/test/expressions.txt.const
+++ b/sys/test/expressions.txt.const
@@ -1,0 +1,3 @@
+arches = 32_fork_shmem, 32_shmem, 64, 64_fork
+FIELD_FLAG1 = 2
+FIELD_FLAG2 = 4