… docs/syscall_descriptions_syntax.md
… pkg/ast/parser_test.go
… pkg/ast/testdata/all.txt
… pkg/ast/testdata/errors.txt
… pkg/compiler/consts.go
… pkg/compiler/consts_test.go
… pkg/compiler/testdata/all.txt
… pkg/compiler/testdata/consts.txt
… pkg/compiler/testdata/errors.txt
… pkg/compiler/testdata/errors2.txt
… pkg/runtest/run_test.go
… prog/encoding.go
… prog/encoding_test.go
… prog/expr.go
… prog/expr_test.go
… prog/minimization.go
… prog/mutation.go
… prog/prog.go
… prog/rand.go
… prog/size.go
… prog/size_test.go
… prog/types.go
… prog/validation.go
… sys/linux/net_80211.txt
… sys/linux/test/80211_ibss
… sys/linux/test/80211_scan
… sys/linux/test/80211_setup_ap
… sys/linux/test/80211_setup_station
… sys/test/exec.txt
… sys/test/expressions.txt