Closes #2226

**WIP**

Support conditional fields with the following syntax:
```
struct {
  mask   int32
  field  some_substruct    (if[value[mask] & SOME_CONST])
  field2 other_substruct   (if[value[mask] & SOME_CONST == 0])  
}
```

In the example above, `field` will only be included if `mask & SOME_CONST != 0`. In the other case, there'll be `field2`.

The code supports however complex expressions that can be made by combining `==`, `!=`, `&` and `value[path:to:field]`.

It should be quite easy to add more operations later. Also, the way it's implemented now will let us add something like `eq[value[other_field]+1]` for cases when we want one field's value to be dependent on other fields' values. If we ever need it, of course.

The underlying implementation relies on conditional fields for unions. But these can be used independently from conditions on struct fields.
```
s  {
  mask   int32
  f1     u
}

u [
  field  some_substruct    (if[value[parent:mask] & SOME_CONST])
  field2 other_substruct   (if[value[parent:mask] & OTHER_CONST])  
  other third_struct
}
```

The union must either have no fields with conditions, or all fields except the last one must have some condition. During fuzzing, syzkaller would pick the first field that matches its conditions (thanks to the limitation in the previous sentence, there will always be a solution).

TODO:
- [x] Add documentation
- [x] Rewrite some descriptions
- [x] Add more `prog` tests