@a-nogikh 2023-12-11
Okay, we actually already use `nil` in `prog/encoding.go`.

For normal arguments, it de facto means "generate the default value":
https://github.com/google/syzkaller/blob/28b24332d95f2f7df44ec7e7a5e0025bcadc6277/prog/encoding.go#L411-L413

We even test for it:
https://github.com/google/syzkaller/blob/28b24332d95f2f7df44ec7e7a5e0025bcadc6277/prog/encoding_test.go#L222-L225

Also the `vm` type can have it as a value (that's when `typ == nil` in `encoding.go`). This use does not contradict the purpose of the PR though.
https://github.com/google/syzkaller/blob/28b24332d95f2f7df44ec7e7a5e0025bcadc6277/prog/encoding_test.go#L323-L324

I'm leaning towards introducing some `EmptyArg` in `prog/prog.go` and using the `empty` keyword to signify this in programs. It will fit more nicely into the existing `prog` design and there won't be tons of `if arg == nil` checks. Progs would look like

```
test$conditional_struct(0x4, empty, 0) 
```