@dvyukov 2023-12-12
Will it be simpler if we lower this to a varlen union with a value and a void field (similar to the current opt template)? Then we can just select the @void field to delete it.

And if we implement unions with conditions on options and expose it in syzland as well, then this will resolve another feature request IIRC (did we want it somewhere?).