@a-nogikh 2023-12-21
> Will it be simpler if we lower this to a varlen union with a value and a void field (similar to the current opt template)?

What exactly do you mean by "lower"?

**A.** If we keep the syntax for structs as initially proposed here
```
A {
    f1   t1
    f2   t2   (if[val[f1] == 0])
} [packed]
```

Then we'd need to somehow create a fake union type in `pkg/compiler` and substitute it as a field type, as if:
```
autogenerated1 [
  f2 t2 (if[val[parent:f1] == 0]) # we don't support `parent:field` now, but let's imagine
] [varlen]

A {
    f1   t1
    f2   autogenerated1
} [packed]
```

* All field references go via `parent:`, which is not supported yet.
* For `prog`, it will be union, so all errors and restrictions (e.g. about resources) will be somewhat confusing.

**B.** Don't support `if[expr]` attr for struct fields (only for union fields). Use conditional fields via a template explicitly.

Built-in:
```
type optField[T, E] [
	val	T (if[E])
	void	void
] [varlen]
```

In descriptions:
```
A {
    f1   t1
    f2   optField[t2, val[parent:f1] == 0]
} [packed]
```

* We still have to carry on `parent:` (and support it).
* The syntax gets somewhat more clumsy.

**Regarding `if[]` for union fields in general:**

We must select exactly one union field, so when we generate a union with field conditions, we must set the values of depend-on fields in such a way that at least one field can be selected. This is doable, but might be quite tricky (especially when we support more expression operators), so we either do that, or we require that one union field remains without a condition.

In the latter case, though it covers exactly the use case we discuss here, it is unlikely to be very useful for anything else.

And even if we support unions with conditions for all fields.. I don't know where exactly we can apply it. If all field conditions are mutually exclusive, it's no different from a struct with field conditions. Otherwise, if conditions of several union fields match, was the use of such a union with field conditions a good choice in the first place?

---

If we support per-field conditions in unions, we can indeed use it to implement conditions on struct fields, but some parts of the implementation will become more tricky than now (take care about `parent:a:b:c` field references and deal with a more complicated generation/mutation).

If we only support per-field conditions in structs, it will reduce the need for field conditions in unions.

---

Did you have any other options in mind?