@dvyukov 2023-12-21
I meant A. Syntax and internal representation don't need to match. We already do a number of such transformations in the compiler.

> For prog, it will be union, so all errors and restrictions (e.g. about resources) will be somewhat confusing.

All errors are printed by the compiler before the transformation.
Don't we want the same restrictions about resources? The field may be missing, so we can't assume that the syscall will create the resource.

> We must select exactly one union field, so when we generate a union with field conditions, we must set the values of depend-on fields in such a way that at least one field can be selected. 

I think we should require a "default" field w/o a condition at the end.
Consequently this should be lowered to:

```
autogenerated1 [
  f2 t2 (if[val[parent:f1] == 0]) # we don't support `parent:field` now, but let's imagine
  v void
] [varlen]
```

> All field references go via parent:, which is not supported yet.

IIRC we support `parent:f1` and `parent:parent:f1` and `A:f1`.