@a-nogikh 2023-12-22
Thanks for the clarification, I'll try to do it this way.

> IIRC we support parent:f1 and parent:parent:f1 and A:f1.

For `parent:f1` -- no, we don't.
https://github.com/google/syzkaller/blob/fb427a0782000106c62de76d251e5a02de5406a9/pkg/compiler/testdata/errors2.txt#L149-L151