@a-nogikh 2023-12-28
I've rewritten the code to use unions underneath. There are still TODOs (including adding more tests, documentation and rewriting some of the existing descriptions).

@dvyukov  could you please take a look?

---

Now `len` also supports the `parent:field` notation, though it's somewhat tricky syntactically now.

For structs, we used to refer to the whole struct as `parent`, so `parent:field` must logically mean a field in the same structure, just as if `parent:` was not there.
```
X {
 f1  array[int32]
 f2  len[parent:f1, int32]
}
```

For unions, there are no sibling fields, so `parent:other-union-field` is meaningless, so `parent` may refer to the wrapping structure. We could save
```
X {
 f1 array[int32]
 f2 Y
}

Y [
 u1 len[parent:f1, int32]
]
```

`prog/size.go` did not initialize `len`-like fields inside unions by the way, even though we used it in the descriptions at least once:

```
type ebt_among_info_offset[FIELD] [
	offset	offsetof[ebt_among_info:FIELD, int32]
	zero	const[0, int32]
]
```

This should be working well after these changes.