@dvyukov 2023-12-29
> For structs, we used to refer to the whole struct as parent, so parent:field must logically mean a field in the same structure, just as if parent: was not there.

Yes, this is nasty. In this case, shouldn't it be `parent:parent:field` to make it consistent?

But also when lowering other constructs, we can just use the parent name explicitly:
```
X {
 f1 array[int32]
 f2 Y
}

Y [
 u1 len[X:f1, int32]
]
```

So perhaps we don't need any extensions for field references for this.