@dvyukov 2023-12-29
Nice!
I think in general this is the right approach and it will provide a good base for implementation of other features on top.

We also need some user docs with details and examples.
In particular, I am interested in the alignment details for struct/union fields. This may also need some additional tests for alignment, I see lots of potential for bugs there.

Are conditional fields supported for bitfields? I think we should explicitly prohibit it for now.

Exact rules for union conditional fields.

Expression types story. == and != produce bool result, while & produce int result. Unclear if the end type of the expression is required to be bool.