@dvyukov 2024-01-03
> I've run some side-to-side comparisons with the current upstream version and the version from this PR seemed ~20% slower (apparently due to extra computations we now perform).
> 
> I've optimized some conditional fields-related code and added the `prog: cache parents map` commit that also affects existing `len[X, Y]` calculations. Now this version seems to be even faster than the upstream one.

Btw, for this type of this things you can use Go's microbenchmarks, e.g.:


```
$ go test -run=none -bench=BenchmarkMutate -benchtime=0.5s -count=10 ./prog | tee /tmp/1
BenchmarkMutate-72    	    5335	     98306 ns/op	  378774 B/op	    3270 allocs/op
BenchmarkMutate-72    	    6972	     93111 ns/op	  423663 B/op	    3320 allocs/op
BenchmarkMutate-72    	    6625	     89595 ns/op	  423379 B/op	    3319 allocs/op
...

$ go test -run=none -bench=BenchmarkMutate -benchtime=0.5s -count=10 ./prog | tee /tmp/2
...

$ benchstat /tmp/1 /tmp/2
          │   /tmp/1    │            /tmp/2             │
          │   sec/op    │   sec/op     vs base          │
Mutate-72   93.67µ ± 1%   97.80µ ± 5%  ~ (p=0.052 n=10)

          │    /tmp/1    │             /tmp/2             │
          │     B/op     │     B/op      vs base          │
Mutate-72   413.6Ki ± 1%   413.2Ki ± 1%  ~ (p=0.165 n=10)

          │   /tmp/1    │               /tmp/2               │
          │  allocs/op  │  allocs/op   vs base               │
Mutate-72   3.325k ± 0%   3.320k ± 0%  -0.15% (p=0.026 n=10)
```