@a-nogikh 2024-01-03
Benchstat:

* `/tmp/upstream` is the current syzkaller HEAD.
* `/tmp/cond-fields` are all base changes from this PR.
* `/tmp/cache-map` is `/tmp/cond-fields` + [prog: cache parents map](https://github.com/a-nogikh/syzkaller/commit/2e5ff016e7660707b452942d01eafc646a508570)
* `/tmp/parent-stack` is `prog: collect parents during arg traversal`

I don't know how suitable that particular benchmark is though. The vast majority of syzkaller executions are minimiziations (and that's presumably where we recalculate conditional fields and lengths most of the time).

```
          │ /tmp/upstream │          /tmp/cond-fields          │            /tmp/cache-map             │          /tmp/parent-stack          │
          │    sec/op     │   sec/op     vs base               │   sec/op     vs base                  │   sec/op     vs base                │
Mutate-36     139.9µ ± 1%   149.6µ ± 5%  +6.89% (p=0.000 n=15)   152.7µ ± 4%  +9.14% (p=0.000 n=15+10)   163.3µ ± 4%  +16.71% (p=0.000 n=15)

          │ /tmp/upstream │           /tmp/cond-fields           │             /tmp/cache-map              │          /tmp/parent-stack           │
          │     B/op      │     B/op      vs base                │     B/op      vs base                   │     B/op      vs base                │
Mutate-36    403.8Ki ± 1%   505.9Ki ± 2%  +25.27% (p=0.000 n=15)   509.0Ki ± 2%  +26.04% (p=0.000 n=15+10)   503.2Ki ± 1%  +24.60% (p=0.000 n=15)

          │ /tmp/upstream │          /tmp/cond-fields          │            /tmp/cache-map             │          /tmp/parent-stack          │
          │   allocs/op   │  allocs/op   vs base               │  allocs/op   vs base                  │  allocs/op   vs base                │
Mutate-36     3.286k ± 0%   3.427k ± 1%  +4.29% (p=0.000 n=15)   3.433k ± 1%  +4.47% (p=0.000 n=15+10)   4.506k ± 1%  +37.13% (p=0.000 n=15)
```