@a-nogikh 2024-01-03
After https://github.com/google/syzkaller/commit/8a70020f31559cb43209df6788c90da90f05cec2 and some performance optimizations of `prog: collect parents during arg traversal` (only collect parents in size.go and expr.go):

```
          │ /tmp/upstream │  /tmp/optimized-parent-slice  │
          │    sec/op     │   sec/op     vs base          │
Mutate-36     117.0µ ± 6%   115.2µ ± 4%  ~ (p=0.653 n=15)

          │ /tmp/upstream │     /tmp/optimized-parent-slice     │
          │     B/op      │     B/op      vs base               │
Mutate-36    403.4Ki ± 1%   375.0Ki ± 2%  -7.05% (p=0.000 n=15)

          │ /tmp/upstream │    /tmp/optimized-parent-slice     │
          │   allocs/op   │  allocs/op   vs base               │
Mutate-36     3.273k ± 0%   3.435k ± 0%  +4.95% (p=0.000 n=15)
```